Refactored handles to be objects which contain a .drop() method.

+++ db/low-level/src/ts/data-id.ts
@@ -10,39 +10,18 @@
  */
 
 import { ApiClient, ApiClientConfig } from "./client";
-import { saneResponse, SafeError, UnexpectedResponseContent } from "./util";
+import { saneResponse, SafeError, UnexpectedResponseContent
+         , InvalidHandleError } from "./util";
 import * as WebRequest from "web-request";
 import { Response, Request } from "web-request";
 import * as stream from "stream";
-
-export type DataIDHandle = number;
+import { Handle } from "./raii";
 
 export class DataIDClient extends ApiClient {
 
     constructor(conf: ApiClientConfig) {
         super(conf);
     }
-
-    /**
-    *
-    * @param handleId - a handleId as obtained by creating and appendable or
-    *                   structured data, and then converting it to a data-id.
-    *                   This can be confusing, so see the tests in
-    *                   `spec/data_id_spec.ts` for an example.
-    * @returns The serialized handleId
-    */
-    public async serialise(handleId: DataIDHandle): Promise<Buffer> {
-        const res: Response<string> =
-            await saneResponse(WebRequest.create<string>(
-            `${this.endpoint}/data-id/${handleId}`, {
-                method: "GET",
-                auth: {
-                    bearer: (await this.authRes).token
-                }
-            }).response);
-
-        return Buffer.from(res.content);
-    };
 
     /**
     *
@@ -77,20 +56,51 @@ export class DataIDClient extends ApiClient {
         if (typeof resObj.handleId === "undefined") {
             throw new UnexpectedResponseContent(response);
         }
-        return resObj.handleId;
+
+        return new DataIDHandle(this, resObj.handleId);
+    }
+
+}
+
+
+export class DataIDHandle extends Handle {
+
+    constructor(c: ApiClient, handle: number) {
+        super(c, handle);
     }
 
     /**
+    *
+    * @returns The serialized handleId
+    */
+    public async serialise(): Promise<Buffer> {
+        if (!this.valid) throw new InvalidHandleError(this.handle);
+
+        const res: Response<string> =
+            await saneResponse(WebRequest.create<string>(
+            `${this.client.endpoint}/data-id/${this.handle}`, {
+                method: "GET",
+                auth: {
+                    bearer: (await this.client.authRes).token
+                }
+            }).response);
+
+        return Buffer.from(res.content);
+    }
+
+
+    /**
      *
-     * @param handle - the structured data handle
      */
-    public async drop(handle: DataIDHandle): Promise<void> {
+    protected async dropImpl(): Promise<void> {
+        if (!this.valid) throw new InvalidHandleError(this.handle);
+
         const result = await saneResponse(WebRequest.create<any>(
-            `${this.endpoint}/data-id/${handle}`, {
+            `${this.client.endpoint}/data-id/${this.handle}`, {
                 method: "DELETE",
                 json: true,
                 auth: {
-                    bearer: (await this.authRes).token
+                    bearer: (await this.client.authRes).token
                 }
             }).response);
         if (result.statusCode !== 200) {
@@ -100,6 +110,7 @@ export class DataIDClient extends ApiClient {
 
 
 }
+
